refactor: remove unused files
