--- src/class_inspector/function_inspector.py
@@ -1,374 +1,0 @@
-import inspect
-import logging
-from typing import Any, Callable, Dict
-
-import attr
-from attr.validators import instance_of
-
-from class_inspector import _type_hint_utils as thu
-from class_inspector import _utils as utils
-from class_inspector._logger import (
-    compress_logging_value,
-    is_logging_enabled,
-    setup_logger,
-)
-
-if is_logging_enabled():
-    setup_logger(__file__, 2)
-
-logger = logging.getLogger()
-
-
-@attr.define
-class FunctionInspector:
-    """A function inspector class for analysing functions and methods as well as
-    writing test stubs
-    """
-
-    obj: Callable = attr.ib(init=False, validator=[instance_of(Callable)])  # type: ignore
-    name: str = attr.ib(init=False, validator=[instance_of(str)])
-    doc: str = attr.ib(init=False, validator=[instance_of(str)])
-    parameters: dict = attr.ib(init=False, validator=[instance_of(dict)])
-    return_annotation: str = attr.ib(init=False, validator=[instance_of(str)])
-    is_method: int = attr.ib(init=False, validator=[instance_of(int)])
-    use_id: bool = attr.ib(default=False, validator=[instance_of(bool)])
-    tab: str = attr.ib(default="    ", validator=instance_of(str), init=False)  # type: ignore
-
-    def analyse(self, object_) -> None:
-        """
-        Analyze the given object to extract information such as name, documentation,
-        parameters, and return annotation.
-
-        Args:
-            object_ (Callable): The callable object to be analysed.
-        """
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-        self.obj = object_
-        self.name = self.obj.__name__
-        self.doc = self._get_doc()
-        self.parameters = self._get_parameters()
-        self.return_annotation = self._get_return_annotations()
-        self.tab: str = "    "
-        # add 1 tab to the start of each line if obj is a method
-        self.is_method = int(inspect.ismethod(self.obj))
-
-    def add_boilerplate(self, add_guards: bool = True, add_debugs: bool = True) -> str:
-        """finds the last index of the doctsring if present and inserts
-        the guard conditions in there
-
-        Returns:
-            str: the analysed function with added guard conditions
-        """
-
-        # replace double quotes with single quotes as strings default to single quotes
-        func_str = utils.clean_func(str(inspect.getsource(self.obj)).replace('"', "'"))
-        end_idx = utils.find_string_end(func_str, utils.get_docstring_patterns())
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-
-        if end_idx:
-            if add_guards:
-                func_str = utils.insert_string_at_idx(
-                    func_str, end_idx, self._get_guards()
-                )
-            if add_debugs:
-                func_str = utils.insert_string_at_idx(
-                    func_str, end_idx, self._get_debugs()
-                )
-        else:
-            sig = self._get_func_sig() + "\n"
-
-            if sig not in func_str:
-                raise ValueError(
-                    "function does not have docstring to locate where to insert boilerplate "
-                    "when using aliased type hints, "
-                    "add docstring to ensure correct behaviour. "
-                    f"sig not found in function str: {sig = } {func_str = }"
-                )
-            if add_guards:
-                func_str = func_str.replace(sig, f"{sig}{self._get_guards()}")
-            if add_debugs:
-                func_str = func_str.replace(sig, f"{sig}{self._get_debugs()}")
-            if self.doc:
-                func_str = func_str.replace(sig, f"{sig}{self.doc}")
-
-            logger.debug(f"sig = {sig}")
-        logger.debug(f"func_str = {func_str}")
-        return utils.format_code_str(func_str)
-
-    def get_test(self, check_types: bool = True, match: bool = False) -> str:
-        """
-        Generate the complete test function.
-
-        Args:
-            check_types (bool): Whether to include type checking in the test
-                function.
-            match (bool): Whether to include match pattern in the test function.
-
-        Returns:
-            str: The complete test function.
-        """
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-        test_full = []
-        test_full.append(self._get_parametrize_decorator(check_types, match))
-        test_full.append(self._get_test_sig())
-        test_full.append(self._get_test_body())
-        test_full.append("\n\n")
-        return utils.format_code_str("".join(test_full))
-
-    def _get_doc(self) -> str:
-        """
-        Get the documentation string of the analysed object.
-
-        Returns:
-            str: The documentation string of the analysed object.
-        """
-        doc = inspect.getdoc(self.obj)
-        if doc:
-            return str(doc)
-        return ""
-
-    def _get_parameters(self) -> Dict[str, Any]:
-        return {
-            param.name: param.annotation
-            for _, param in inspect.signature(self.obj).parameters.items()
-        }
-
-    def _get_return_annotations(self) -> str:
-        """
-        Get the return annotation of the analysed object.
-
-        Returns:
-            str: The return annotation of the analysed object.
-        """
-        annot = inspect.signature(self.obj).return_annotation
-        if annot is not inspect._empty and not thu.is_union_origin(annot):
-            return thu.get_object_name(annot)
-        if thu.is_union_origin(annot):
-            return str(annot).replace("typing.", "")
-        return "None"
-
-    def _get_parametrize_decorator(
-        self, check_types: bool = True, match: bool = False
-    ) -> str:
-        """
-        Generate a parametrize decorator for the test function.
-
-        Args:
-            check_types (bool): Whether to include type checking in the test cases.
-            match (bool): Whether to include match pattern in the test cases.
-
-        Returns:
-            str: The parametrize decorator for the test function.
-        """
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-        args = self._get_params_str()
-        return "".join(
-            [
-                "@pytest.mark.parametrize(\n",
-                f'{self.tab}"{args}, expected_result, expected_context",\n',
-                f"{self.tab}[\n",
-                self._get_test_case(args),
-                (
-                    self._get_raises_type_error_test_case(args, check_types, match)
-                    * len(self.parameters)
-                ),
-                f"{self.tab}]\n)\n",
-            ]
-        )
-
-    def _get_test_sig(self) -> str:
-        """
-        Get the signature string for generating a test function.
-
-        Returns:
-            str: The signature string for generating a test function.
-        """
-        sig = self._get_instance_sig() + self._get_params_str()
-        return (
-            f"def test_{utils.strip_underscores(self.name)}"
-            f"({sig}, expected_result, expected_context) -> None:\n"
-        )
-
-    def _get_test_body(self) -> str:
-        """
-        Generate the body of the test function.
-
-        Returns:
-            str: The body of the test function.
-        """
-        test_body = [f"{self.tab}with expected_context:\n"]
-        test_body.append(f"{2*self.tab}assert ")
-        test_body.append(self._get_instance_call())
-        if self.return_annotation != "None":
-            test_body.append("== expected_result\n")
-        else:
-            test_body.append("is None\n")
-        return "".join(test_body)
-
-    def _get_params_str(self) -> str:
-        """
-        Get a string representation of the parameters of the analysed object.
-
-        Returns:
-            str: A string representation of the parameters of the analysed object.
-        """
-        params = ", ".join(list(self.parameters.keys()))
-        if params:
-            return params
-        return ""
-
-    def _get_test_case(self, args: str) -> str:
-        """
-        Generate a test case for the test function.
-
-        Args:
-            args (str): The string representation of function arguments.
-
-        Returns:
-            str: The test case for the test function.
-        """
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-        if self.use_id:
-            test_id = ', id=""'
-        else:
-            test_id = ""
-        return f"{self.tab * 2}pytest.param({args}, expected_result, expected_context{test_id}),\n"
-
-    def _get_raises_type_error_test_case(
-        self, args: str, check_types: bool = True, match: bool = False
-    ) -> str:
-        """
-        Generate a test case for raising a type error in the test function.
-
-        Args:
-            args (str): The string representation of function arguments.
-            check_types (bool): Whether to include type checking in the test cases.
-            match (bool): Whether to include match pattern in the test cases.
-
-        Returns:
-            str: The test case for raising a type error in the test function.
-        """
-        logger.debug(
-            {key: compress_logging_value(val) for key, val in locals().items()}
-        )
-        if not check_types:
-            return ""
-        match_stmt = ""
-        if match:
-            match_stmt = ', match=""'
-        return f"{self.tab * 2}pytest.param({args}, None, pytest.raises(TypeError{match_stmt})),\n"
-
-    def _get_instance_sig(self) -> str:
-        """
-        Get the signature string for calling an instance method (if applicable).
-
-        Returns:
-            str: The signature string for calling an instance method.
-        """
-        if inspect.ismethod(self.obj):
-            instance = utils.camel_to_snake(self._get_class_name())
-            return f"get_{instance}_instance: {self._get_class_name()}, "
-        if inspect.isfunction(self.obj):
-            return ""
-        return ""
-
-    def _get_instance_call(self) -> str:
-        """
-        Generate a string representation for calling the analysed object.
-
-        Returns:
-            str: A string representation for calling the analysed object.
-        """
-        sig: str = self._get_params_str()
-        if inspect.ismethod(self.obj):
-            instance = utils.camel_to_snake(self._get_class_name())
-            return f"get_{instance}_instance.{self.name}({sig}) "
-        if inspect.isfunction(self.obj):
-            return f"{self.name}({sig}) "
-        return f"{self.name}({sig}) "
-
-    def _get_class_name(self) -> str:
-        """
-        Get the class name of the analysed object (if it's a method).
-
-        Returns:
-            str: The class name of the analysed object.
-        """
-        if inspect.ismethod(self.obj):
-            return self.obj.__self__.__class__.__name__
-        return self.name
-
-    def _get_func_sig(self) -> str:
-        """returns the function signature
-        supports methods by inserting self as the first arg
-
-        Returns:
-            str: function signature
-        """
-        definition = "def "
-        init_sig = str(inspect.signature(self.obj))
-
-        if inspect.ismethod(self.obj):
-            definition = definition.replace("def ", "    def ")
-            init_sig = init_sig.replace("(", "(self, ").replace(", )", ")")
-
-        sig = f"{definition}{self.name}" + init_sig + ":"
-        return sig
-
-    def _get_guards(self) -> str:
-        """
-        Generate guards for type checking of function arguments.
-
-        Returns:
-            str: Guards for type checking of function arguments.
-        """
-        if not self.parameters:
-            return ""
-
-        expected_types, received_types = [], []
-
-        for arg, annot in self.parameters.items():
-            if annot != inspect._empty:
-                expected_types.append(thu.unpack_parameter(annot))
-                received_types.append(f"{{type({arg}).__name__}}")
-
-        expected_types = ", ".join(expected_types)
-        received_types = ", ".join(received_types)
-
-        if expected_types and received_types:
-            guards: str = (
-                f"{self.tab*(1 + self.is_method)}if not all(["
-                + ", ".join(
-                    [
-                        f"isinstance({arg_name}, {thu.unpack_parameter(arg_type)})"
-                        for arg_name, arg_type in self.parameters.items()
-                        if arg_type != inspect._empty
-                    ]
-                )
-                + "]):\n"
-            )
-            raises = (
-                f"{self.tab*(2 + self.is_method)}raise TypeError(\n"
-                + f'{self.tab*(3 + self.is_method)}"{self.name} '
-                + f'expects arg types: [{expected_types}], "\n'
-                + f'{self.tab*(3 + self.is_method)}f"received: [{received_types}]"\n'
-                + f"{self.tab*(2 + self.is_method)})\n"
-            )
-            return guards + raises
-        return ""
-
-    def _get_debugs(self) -> str:
-        if not self.parameters:
-            return ""
-        debugs = f"{self.tab*(1 + self.is_method)}logger.debug(locals())\n"
-        return debugs
